Add support for omitting the `localPath` in a `git.clone` Update docs accordingly.
Closes #196
