+++ readme.md
@@ -39,7 +39,7 @@ when no error occurred) and secondly the data generated by that call.
 | `.checkoutLocalBranch(branchName, handlerFn)` | checks out a new local branch |
 | `.deleteLocalBranch(branchName, handlerFn)` | deletes a local branch |
 | `.checkoutLatestTag(handlerFn)` | convenience method to pull then checkout the latest tag |
-| `.clone(repoPath, localPath, [options], [handlerFn])` | clone a remote repo at `repoPath` to a local directory at `localPath` with an optional array of additional arguments to include between `git clone` and the trailing `repo local` arguments |
+| `.clone(repoPath, [localPath, [options]], [handlerFn])` | clone a remote repo at `repoPath` to a local directory at `localPath` (can be omitted to use the default of a directory with the same name as the repo name) with an optional array of additional arguments to include between `git clone` and the trailing `repo local` arguments |
 | `.clearQueue()` | immediately clears the queue of pending tasks (note: any command currently in progress will still call its completion callback) |
 | `.commit(message, handlerFn)` | commits changes in the current working directory with the supplied message where the message can be either a single string or array of strings to be passed as separate arguments (the `git` command line interface converts these to be separated by double line breaks) |
 | `.commit(message, [fileA, ...], options, handlerFn)` | commits changes on the named files with the supplied message, when supplied, the optional options object can contain any other parameters to pass to the commit command, setting the value of the property to be a string will add `name=value` to the command string, setting any other type of value will result in just the key from the object being passed (ie: just `name`), an example of setting the author is below |
