+++ test/unit/test-commands.js
@@ -42,44 +42,6 @@ exports.childProcess = {
     }
 };
 
-
-exports.clone = {
-    setUp: function (done) {
-        git = Instance();
-        done();
-    },
-
-    'clone with repo and local': function (test) {
-        git.clone('repo', 'lcl', function (err, data) {
-            test.same(['clone', 'repo', 'lcl'], theCommandRun());
-            test.same('anything', data);
-            test.equals(null, err, 'not an error');
-
-            test.done();
-        });
-
-        closeWith('anything');
-    },
-
-    'clone with options': function (test) {
-        git.clone('repo', 'lcl', ['foo', 'bar'], function (err, data) {
-            test.same(['clone', 'foo', 'bar', 'repo', 'lcl'], theCommandRun());
-            test.done();
-        });
-
-        closeWith('anything');
-    },
-
-    'explicit mirror': function (test) {
-        git.mirror('r', 'l', function () {
-            test.same(['clone', '--mirror', 'r', 'l'], theCommandRun());
-            test.done();
-        });
-
-        closeWith();
-    }
-};
-
 exports.diff = {
     setUp: function (done) {
         git = Instance();
